refactor: move ZoomControls to components/ui
