+++ web/src/components/ui/ZoomControls.tsx
